+++ src/Mapping/FieldMapping.php
@@ -23,6 +23,9 @@ namespace MacFJA\RediSearch\Integration\Mapping;
 
 use function array_key_exists;
 use DOMAttr;
+use Error;
+use Exception;
+use function get_class;
 use InvalidArgumentException;
 use function is_string;
 use MacFJA\RediSearch\Redis\Command\CreateCommand\CreateCommandFieldOption;
@@ -30,6 +33,9 @@ use MacFJA\RediSearch\Redis\Command\CreateCommand\GeoFieldOption;
 use MacFJA\RediSearch\Redis\Command\CreateCommand\NumericFieldOption;
 use MacFJA\RediSearch\Redis\Command\CreateCommand\TagFieldOption;
 use MacFJA\RediSearch\Redis\Command\CreateCommand\TextFieldOption;
+use ReflectionException;
+use ReflectionMethod;
+use ReflectionProperty;
 
 trait FieldMapping
 {
@@ -38,17 +44,68 @@ trait FieldMapping
      */
     protected function getFieldValue(object $object, ?string $getter, ?string $property, ?string $fieldName)
     {
-        if (is_string($getter) && method_exists($object, $getter)) {
-            return $object->{$getter}();
+        if (is_string($getter)) {
+            try {
+                return $this->getObjectMethod($object, $getter);
+            } catch (Exception $exception) {
+                // Do nothing
+            }
         }
-        if (is_string($property) && property_exists($object, $property)) {
-            return $object->{$property};
-        }
-        if (is_string($fieldName) && property_exists($object, $fieldName)) {
-            return $object->{$fieldName};
+        $name = $property ?? $fieldName;
+        if (is_string($name)) {
+            try {
+                return $this->getObjectProperty($object, $name);
+            } catch (Exception $exception) {
+                // Do nothing
+            }
         }
 
         return null;
+    }
+
+    /**
+     * @return mixed
+     */
+    protected function getObjectProperty(object $object, string $property)
+    {
+        $data = get_object_vars($object);
+        if (array_key_exists($property, $data)) {
+            return $data[$property];
+        }
+
+        try {
+            $propertyReflection = new ReflectionProperty($object, $property);
+            $propertyReflection->setAccessible(true);
+
+            return $propertyReflection->getValue($object);
+        } catch (ReflectionException $exception) {
+            // Do nothing
+        }
+
+        throw new InvalidArgumentException(sprintf('Unable to read property %s from %s', $property, get_class($object)));
+    }
+
+    /**
+     * @return mixed
+     */
+    protected function getObjectMethod(object $object, string $method)
+    {
+        try {
+            return $object->{$method}();
+        } catch (Error $exception) {
+            // Do nothing
+        }
+
+        try {
+            $propertyReflection = new ReflectionMethod($object, $method);
+            $propertyReflection->setAccessible(true);
+
+            return $propertyReflection->invoke($object);
+        } catch (ReflectionException $exception) {
+            // Do nothing
+        }
+
+        throw new InvalidArgumentException(sprintf('Unable to read method %s from %s', $method, get_class($object)));
     }
 
     /**
@@ -84,12 +141,16 @@ trait FieldMapping
         }
         $option->setField($name);
         $option = $this->setOptionValue($option, $options, 'sortable', 'setSortable', CreateCommandFieldOption::class, 'bool');
+        $option = $this->setOptionValue($option, $options, 'unnormalized', 'setUnNormalizedSortable', CreateCommandFieldOption::class, 'bool');
         $option = $this->setOptionValue($option, $options, 'noindex', 'setNoIndex', CreateCommandFieldOption::class, 'bool');
+
         $option = $this->setOptionValue($option, $options, 'phonetic', 'setPhonetic', TextFieldOption::class, 'string');
         $option = $this->setOptionValue($option, $options, 'nostem', 'setNoStem', TextFieldOption::class, 'bool');
         $option = $this->setOptionValue($option, $options, 'weight', 'setWeight', TextFieldOption::class, 'float');
 
-        return $this->setOptionValue($option, $options, 'separator', 'setSeparator', TagFieldOption::class, 'string');
+        $option = $this->setOptionValue($option, $options, 'separator', 'setSeparator', TagFieldOption::class, 'string');
+
+        return $this->setOptionValue($option, $options, 'casesensitive', 'setCaseSensitive', TagFieldOption::class, 'bool');
     }
 
     /**
@@ -105,22 +166,28 @@ trait FieldMapping
     }
 
     /**
+     * @param bool|DOMAttr|float|string $value
+     *
      * @return bool|float|string
      */
-    private function castAttribute(DOMAttr $value, string $cast)
+    private function castAttribute($value, string $cast)
     {
+        if ($value instanceof DOMAttr) {
+            $value = $value->textContent;
+        }
+
         switch ($cast) {
             case 'bool':
             case 'boolean':
-                return 'true' === $value->textContent;
+                return 'true' === $value || true === $value;
 
             case 'float':
             case 'double':
-                return (float) $value->textContent;
+                return (float) $value;
 
             case 'string':
             default:
-                return $value->textContent;
+                return (string) $value;
         }
     }
 }