+++ CHANGELOG.md
@@ -6,6 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add new configuration (UnNormalizedForm, CaseSensitive) from Xml and Json mapping
+
+### Fixed
+
+- Fix tag not transformed from array to string
+- Fix PHP8 attribute not read correctly
+- Fix Json mapping not working because of Xml mapping
+- Fix add document to search not working (empty document in Redis)
+- (dev) Coverage with XDebug 3
+
+### Changed
+
+- (dev) Add compatibility with PHP for QA tools
+- Set the minimum version of `macfja/redisearch` to `2.1.0`
+
 ## [2.0.0]
 
 Rework to be compatible with `macfja/redisearch`=`^2.0`